Updated examples
- Replaced size_t with std::size_t or GLuint in examples

+++ example/oglplus/002_shader_lit.cpp
@@ -94,7 +94,7 @@ public:
 		gl.ClearDepth(1.0f);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 	}
