+++ example/oglplus/004_newton.cpp
@@ -138,7 +138,7 @@ public:
 		gl.Disable(Capability::DepthTest);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 	}
