+++ example/oglplus/004_rect.cpp
@@ -115,7 +115,7 @@ public:
 		gl.Disable(Capability::DepthTest);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 	}
