+++ example/oglplus/018_newton_landscape.cpp
@@ -21,7 +21,7 @@ namespace oglplus {
 class LandscapeExample : public Example
 {
 private:
-	size_t grid_side;
+	GLuint grid_side;
 
 	// helper object building plane vertex attributes
 	shapes::Plane make_plane;
@@ -214,7 +214,7 @@ public:
 		gl.CullFace(Face::Back);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 		prog.Use();
