+++ example/oglplus/019_helium.cpp
@@ -55,7 +55,7 @@ public:
 		// bind the VAO for the sphere
 		sphere.Bind();
 
-		const size_t n_attr = 2;
+		const GLuint n_attr = 2;
 		// pointers to the vertex attribute data build functions
 		typedef GLuint (shapes::Sphere::*Func)(std::vector<GLfloat>&) const;
 		Func func[n_attr] = {
@@ -67,7 +67,7 @@ public:
 		// vertex attribute identifiers from the shaders
 		const GLchar* ident[n_attr] = {"Position", "Normal"};
 
-		for(size_t i=0; i!=n_attr; ++i)
+		for(GLuint i=0; i!=n_attr; ++i)
 		{
 			// bind the VBO
 			vbo[i].Bind(Buffer::Target::Array);
@@ -250,7 +250,7 @@ public:
 		gl.Enable(Capability::DepthTest);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 		auto projection = CamMatrixf::PerspectiveX(
