+++ example/oglplus/024_simple_picking.cpp
@@ -228,7 +228,7 @@ public:
 		gl.Enable(Capability::CullFace);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 		auto perspective = CamMatrixf::PerspectiveX(
