+++ example/oglplus/025_subroutines.cpp
@@ -281,7 +281,7 @@ public:
 	{
 		// setup the texture
 		{
-			size_t tex_side = 512;
+			GLuint tex_side = 512;
 			auto image = images::NewtonFractal(
 				tex_side, tex_side,
 				Vec3f(0.2f, 0.1f, 0.4f),
@@ -311,7 +311,7 @@ public:
 		gl.CullFace(Face::Back);
 	}
 
-	void Reshape(size_t width, size_t height)
+	void Reshape(GLuint width, GLuint height)
 	{
 		gl.Viewport(width, height);
 		cube_prog.projection_matrix = CamMatrixf::PerspectiveX(
